server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
pwdlib[argon2]==0.2.1
email-validator==2.2.0
python-multipart==0.0.20
